radial chart

+++ ExoPlanetHunter.Service/Interfaces/IChartService.cs
@@ -12,6 +12,6 @@ namespace ExoPlanetHunter.Service.Interfaces
     {
         IQueryable<HertzsprungRussellDto> GetHertzsprungRussell(bool habitableOnly);
         IQueryable<IGrouping<string, Planet>> GetPlanetTypes(ChartType type);
-        IQueryable<PlanetDistanceDto> GetPlanetDistance(int? max);
+        IQueryable<PlanetDistanceDto> GetPlanetDistance(double? max);
     }
 }
